requirements.txt
google-cloud-aiplatform
google-auth
google-auth-oauthlib
flask
python-dotenv
